Add:  a coroutine called async_comprehension

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -11,7 +11,7 @@ from typing import Generator
 import asyncio
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     '''Returns a random float number.
     '''
     for _ in range(10):
